script/requirements.txt
charset-normalizer==3.4.4
colorama==0.4.6
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pdf2image==1.17.0
pillow==12.0.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
reportlab==4.4.4
six==1.17.0
tqdm==4.67.1
tzdata==2025.2
